Fix bug where closing of customizer notification notice in WP 4.9 wasn't working properly.
Upcoming WordPress 4.9 display a notice when there is an unsaved/unpublished autosave revision. This lib causes conflict by leaving a top margine of 119px.
This is as a result of this lib dismiss js handler running first before that of customizer.
To fix this issue, I had to prevent this lib dismiss.js code from running in WordPress customizer.